fix letter case inconsistency and more objects
This mainly fixes #13. Since i've wrote a script to help me with that and did lots of find+replace on my files i figured i might as well update all the objects i'm working on on the current state they are.

Still the changelog will only be updated when i get 0.3.0 ready.